Modify bodhi.server.models.BodhiBase.get() to not require db.
Bodhi has a lot of silly code that passes around a database session
which is not necessary. While working on #2297 today I about
mentally snapped on one of these so decided to go crazy and fix
this one. There are many more to fix, but this patch makes the
world a better place.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>