Do not require composer configs to be paths that exist in server.
The goal of this patch is to allow us to split the Bodhi composer
out of the Bodhi server package so we can avoid installing Pungi on
the web server (or non-composer backends). This patch reworks the
configuration verification code so that it only ensures that
composer related path settings are strings, and now the composer
itself ensures the settings refer to existing paths upon startup.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>