Remove a conflicting HTTPForbidden handler.
Beginning with Cornice 3.2.0, conflicting Pyramid configurations
will raise an Exception during application startup. Bodhi had two
views configured to handle HTTPForbidden - one was the
bodhi.server.views.generic.exception_view(), and the other was
bodhi.server.security.login(). This commit removes the routing to
the login() call in favor of the exception_view() call, as logged
in users should not get redirected to the login view, and the code
does not consider whether the user is currently logged in.
Furthermore, this choice is consistent with how Bodhi behaved with
Cornice < 3.2.0.

As a result, Bodhi is now compatible with the latest release of
Cornice.

fixes #2258

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>