Do not apply Koji tags to ejected updates.
Due to a combination of two bugs, Bodhi's composer attempted to
tag ejected updates' builds with the tag None, which is of course
not a valid tag. This happened due to two bugs. The first was that
the ComposerThread did not refresh its self.compose.updates
attribute after ejecting updates, which meant that ejected updates
were still present in operations it peformed after ejecting them.
Secondly, the Update model has a property that returns the koji tag
that correspond's with the Update's request. Since ejected updates
no longer have a request, this method was returning None and then
the composer was handing that value to the Koji client to tag the
build. The property has been altered to raise a RuntimeError if it
is asked for a Koji tag when there is no request.

In addition to the above, this commit also adjusts the DevBuildsys
such that it will raise an Exception if any build is tagged with
None to help prevent this in the future.

fixes #2368

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>