Ensure that the web app closes db connections when done with them.
Prior to this commit, there was a bug in Bodhi where requests that
did not access the request's db attribute would not close out their
database connection. This happened notably with the /composes/ API,
which only used the scoped session to perform its queries.

The commit refactors Bodhi to use Pyramid's event system to attach
a callback to all requests so that the database session is
committed or rolled back (depending on errors) no matter which way
the database session is acquired, so long as it is still the scoped
session. This approach is more general, and works for code that
uses the old or new styles of accessing database session.

As a result of this refactor, the get_db_session_for_request()
function no longer involves itself in transaction management and
simply returns a scoped database session.

fixes #2385

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>