Rename Vagrantfile.example to Vagrantfile.
@abitrolly recently sent a PR[0] to document the
VAGRANT_VAGRANTFILE environment variable, and through discussion
there I was convinced that it was better to have a Vagrantfile
checked in to the root of the repository. This commit makes that
change, and it documents how developers can use a custom
Vagrantfile if they so choose.

This commit was based on the change[0] submitted by @abitrolly.

[0] https://github.com/fedora-infra/bodhi/pull/2262

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>