validators.py: avoid spurious error messages after schema validation fails
If validation against the colander schema for the request fails, then
request.validated is left empty - this will cause subsequent validation
to fail for reasons that have nothing to do with what was submitted and
returning the error messages to the user can be quite confusing.

For example: if a new update is submitted with empty notes, not only does
the user see 'Notes : Required', they also see 'Builds : You may not specify
an empty list of builds.' and 'Builds : ACL validation mechanism was unable
to determine ACLs.' - which are entirely spurious.

In all validators that are run after the colander schema validator, check
to see if a) request.validated is empty b) request.errors is non-empty,
and in this case skip further validation. This is done as a decorator to
have to avoid having dozens of separate code paths for validation-after-schema-
failure that would need to be tested/covered.

Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>