Make Release.from_tags() more predictable
If Release.from_tags() was called with a tag not associated with any
release, it would raise KeyError (unless an earlier tag in the list
was found first.) This was unexpected from the documentation - and was
handled by one caller but not others. Change it to return None in
this case - this was handled by all callers.

Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>