Install python2-backports-functools_lru_cache in Rawhide CI.
This commit is a temporary workaround for #1607916[0]. Once it is
merged I will open a new pull request to undo it. This should allow
CI to pass until the fix for that issue propagates to the mirrors.

[0] https://bugzilla.redhat.com/show_bug.cgi?id=1607916

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>
(cherry picked from commit e084a42a1330a88529e72ac814c7670630c0a167)