Remove tools/signable.py.
tools/signable.py still uses the old lock files that have not
existed in more than 8 months, and thus can not possibly work. As
nobody has complained about this, I have decided to remove the
utility as it does not pass Python 3 flake8 tests and I don't
believe it is worth the effort to maintain.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>