bodhi.client.edit() can assume the update has been validated.
This commit refactors bodhi.client.edit() to rely on
_validate_edit_update() to ensure that the update is either an ID
or an nvr. This resolves an uncovered branch in the test suite,
as formerly it again ensured that the update was either given by ID
or nvr, and there was no way for the test suite to bypass those two
conditions in the branch since it was guarded by the check.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>