Do not marked Duffy nodes as failed.
In recent weeks, Bodhi started using far more test jobs per pull
request in order to give more granular test reporting. When jobs
failed, Bodhi's CI system would mark the nodes used as failed,
which gave use 12 hours to log into those systems to poke around.
This was a nice feature, but now that we run 20 tests per pull
requests, I've been informed that we are now causing large numbers
of systems in CentOS CI to wait around for 12 hours in case we want
to debug on them. I rarely use that feature (though it is nice), so
this pull requests adjusts the script such that we always return
the node as successful so we don't exhaust the pool of workers.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>