Increase the size of the Vagrant guest.
With the recent addition of the bci alias, the Vagrant guest would
benefit from having more RAM and virtual CPUs. bci all will fail
with the current allocation of RAM, and due to the large number of
parallel processes it is beneficial to give the guest more virtual
cores as well.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>