Temporarily disable JJB CI.
I am working on adding a Jenkinsfile[0] to test Bodhi with the
Jenkins Pipeline plugin, and it is difficult to test the new script
when the existing JJB tasks are taking up all the available
workers. This commit temporarily disables the JJB jobs so I can
iterate more rapidly on replacing them.

[0] https://github.com/fedora-infra/bodhi/pull/2609

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>