Allow bodhi-ci tests to wait on builds.
Prior to this commit, bodhi-ci would run all the builds in
parallel, but would wait for all the builds to finish before
starting any of the tests that depended on them. This was
inefficient, and could also be annoying when certain builds fail
for reasons beyond our control (as often happens with Rawhide).

Now it is able to mark test jobs as depending on other test jobs,
so that they can wait for their dependent job to finish before
starting.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>