Add the bodhi.messages Python package
fedora-messages uses Python classes to allow authors to define JSON
schema and a Python API to work with their messages. This adds the
basics of a Python package to distribute these schema.

This is similar to fedora_infrastructure_fedmsg_meta, except each
application manages its own schema definitions and how the message is
represented to the user.

Signed-off-by: Jeremy Cline <jcline@redhat.com>
Signed-off-by: Aurélien Bompard <aurelien@bompard.org>
Signed-off-by: Randy Barlow <randy@electronsweatshop.com>