Add integration testing to bodhi-ci
It currently only works for F29 and Python 3 (which is what we currently
have in prod) but it should be possible to expand that in the future if
there's a need.

The created container is different from the currently created bodhi
containers because it tries to be as close as possible to the production
setup: bodhi is running behind apache, with the production.ini
configuration file, etc.

The integration test suite executes on the host running bodhi-ci via
pytest, but will start the components in containers and test Bodhi
there.

All the CLI's read operations are tested for now.

Signed-off-by: Aurélien Bompard <aurelien@bompard.org>