Handle Koji returning a single timestamp for old changelogs.
Koji seems to return a list of timestamps, even when there is only
one, when retrieving changelog data. Bodhi had been using element
0 of the list if there were more than 1 item in it, but would not
do so if there was exactly one item. This caused an exception
later in some code that was expecting to receive an int instead of
a list with one int inside.

fixes #2768

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>