Allow CLI users to override the OpenID API.
It has long bothered me that Bodhi development uses Fedora's
production authentication server instead of some other
authentication, because this adds a risk that actions in
development might alter production Bodhi (particular with the CLI).

This commit makes it possible for CLI users to override the OpenID
API used by the CLI, and configures the defaults in Vagrant to use
Fedora's staging identity server instead of production.

I did consider giving the server an API to provide settings like
the OpenID provider, but that would cause another request on every
command and so I opted for allowing the user to specify an OpenID
API.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>