Assert update title instead of build nvr in CLI test output.
There was an integration test asserting for the phrase
"Downloading packages from {Build.nvr}" for each Build, but the
CLI was actually printing Update titles. Multi-build updates have
titles that list all the builds, so this test was occasionally
failing, depending on which build was first in the title.

This patch asserts that the Update title is found there instead of
Build NVRs.

fixes #2835

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>