Define a new bugzilla_api_key setting.
This setting adds a new preffered way for Bodhi to authenticate to
Bugzilla when using the XML-RPC API to comment on and change the
status of bugs. It is hoped that this will fix an issue that Bodhi
has been having since Red Hat's Bugzilla instance was upgraded to
version 5 where Bodhi frequently gets an error message stating that
it needs to log in.

fixes #2827

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>