Prepare for the Bodhi 4.0.0 release.
Normally I make release preparation commits when I'm making a
release, but 4.0.0 is expected to have a large number of backwards
incompatible changes, and I would like each of those commits to
include an entry in the changelog, rather than writing the
changelog entries "posthumously" as I usually do. This commit adds
a section to the release notes for 4.0.0, and it also trims the
changelog by moving the old 3.x and 2.x releases into their own
pages. It also sets the version to 4.0.0.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>