Drop support for Python 2 in bodhi.server.
Python 2 support goes away upstream in two months, so it is time
to embrace the glorious future.

This change also drops support for pkgdb integration, since the
pkgdb client library does not support Python 3.

Mock imports were moved to be grouped with the stdlib imports as
per PEP-8, since mock is part of the stdlib in Python 3.

fixes #1970
fixes #2759

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>