Compare strings to strings to in Greenwave decision in web UI.
There was a bug discovered by dmoluguw and investigated by Adam
Williamson where the web interface always used the stable decision
context for updates, even when they were pending testing. The bug
was due to an if statement that compared Enums to strings in the
HTML template, rather than comparing the the Enums directly.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>