Do not alter global settings in test_renderers.py.
While working on #2241, I discovered that removing the stacks
tests resulted in test failures in comment service tests. Via a
bisection process, I discovered that test_renderers.py had been
altering the global configuration in a way that presumably had
been repaired by the stacks tests that I was removing. This commit
fixes test_renderers.py to use mock to alter the settings so the
alterations do not persist after the test is done.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>