Fetch all branches during CI.
We recently altered the Jenkins config so that it only tests PRs
as they are, rather than testing them merged with their destination
branch. This apparently caused the git checkout not to have the
reference for the destination branch, which is needed for
diff-cover to work.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>
(cherry picked from commit 0d8c4915df61579329200e3819d810581e55c5a0)