Add a fake Pungi that can be used during development.
Since Bodhi's development environment does not have a running Koji
yet, it is not possible to run Pungi to test changes to the
composer. This commit introduces a fake Pungi that creates enough
of a repository that the compose can succeed locally. It does not
really compose the updates, it just does the minimum needed to
trick Bodhi into thinking the compose suceeded.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>