Limit how many concurrent integration tests run.
I found that on my laptop, the integration tests would often fail
when run with bodhi-ci's default -j flag in a Vagrant guest. I
discovered that the failures were very often due to timeouts in
operations on the postgresql database, and I also discovered that
running bodhi-ci with a small -j flag (such as -j 2) would get them
to pass. Rather than defaulting to a smaller -j flag value, this
commit adds a second concurrency semaphore for IntegrationJobs to
limit how many of them can run at once.

The chosen value is admittedly a bit "magic", but was chosen
because it happens to get the tests to pass on my Lenovo T480s with
a Vagrant guest.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>