Catch an APIError when killing Greenwave.
There seems to be a frequent issue with Greenwave not working in
some way during integration testing. It seems that it may not be
starting, and one of our integration tests fails to get the IPv4
address for the Greenwave container.

When the test fails, the cleanup code runs, which tries to kill
the Greenwave container. Since the Greenwave container did not
seem to start in the first place, *this* code then fails and
produces a traceback. For some reason, this second traceback seems
to cover up the original problem, which made discovering the issue
difficult.

This commit catches the second Exception and passes. Since we
really just want to ensure that the container is stopped, it's OK
if asking it to stop raises an Exception. We can ignore it, since
that Exception means we are in the desired state anyway.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>