Print Greenwave logs if it fails to start.
There was an issue[0] with the CI tests that was difficult to
debug because the Greenwave logs weren't being captured by the test
suite when it failed to start. This commit adds a startup timeout
to the Greenwave container, and prints the logs from Greenwave if
it fails to start by the timeout.

[0] https://github.com/fedora-infra/bodhi/issues/2976

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>