Skip downloading database dumps when reasonable.
Prior to this commit, every run of the integration test suite would
download a copy of the Fedora Infrastructure's database dumps. The
infrastructure only produces a backup once a day, so this is
inefficient for developers who work on Bodhi often.

This commit adjusts it to only download the database dump if we
don't have it already or if the copy we have is older than a day.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>