Add a new integration-download command to bodhi-ci.
Prior to this commit, if several bodhi-ci processes were run in
parallel, each of them would download their own copy of the
database dump. Currently this is how CentOS CI runs our tests, and
this led to a problem where one process would overrwrite the file
that another process was currently reading, which led to problems.

This new command allows us to run a single process of bodhi-ci to
download the integration test datatbase dumps before we fan out to
many processes in CentOS CI. Combined with the prior commit that
skips downloading recently modified files, we can now be sure that
the fanned out processes in CentOS CI will not write files that
others are reading, because they will all skip the download job
now.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>