Use a UUID in the test network name.
When run in parallel, the integration test processes all created
the same network name. This caused trouble when the tests ended,
because they would try to remove the network that the others were
still using.

This commit gives the test network a UUID so it has a unique name.

fixes #2983

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>