Avoid deadlock when database downloads are old.
I recently wrote a patch for bodhi-ci that attempted to skip
downloading database dumps if they were recently downloaded.
Unfortunately, my implementation's control logic caused a deadlock
if the download existed locally, but was older than a day because
the superclass's run() was never called and the function returned.
This caused a deadlock.

This patch ensures that we call the superclass run() when we have
a copy of the database that is older than one day.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>