Use a setting to configure the fedmsg-hub WS URL.
The fedmsg-hub web socket URL was hardcoded in live.js. This commit
adds a setting for that URL and injects it into the master template
via a <script> tag so that all subsequently loaded JavaScript can
assume that setting is available for use.

This change also adds a nice place for other settings to be added
later, if desired.

fixes #2523

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>