Rename prefer_ssl to libravatar_prefer_tls.
The prefer_ssl setting is renamed to libravatar_prefer_tls by this
commit. The previous name was far too generic of a name, and
implied that it had some global impact when it only affected
libravatar.

The new setting also defaults to True instead of None, a saner
default.

fixes #1921

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>