Update fedora-bootstrap to latest 1.5.0
Update fedora-bootstrap to latest version 1.5.0.
also removed old versions of bootstrap no longer used

This commit was originally several, so their summaries
and body messages are preserved below:

change padding and margin classes to correct syntax

Previous versions of Fedora Bootstrap used an alpha of
Bootstrap 4, which used a different notation for classes
that changed the padding or margin for an element. This
commit updates these classes to the stable bootstrap 4
style.

Make navbar not stack at large sizes

When using navbars with the stable versions of Bootstrap 4,
the navbar needs to specify the nav is horizontal with the
'navbar-expand' class. This commit adds this class to the
main top level menu.

update class syntax for pull/float

The previous version of fedora bootstrap used the pull-*
classes for floating elements. This updates to the float-*
syntax for the newest version of fedora bootstrap

update class syntax for card-block/card-body

The previous version of fedora bootstrap used the card-block
class for the body element of a card element. This updates these
to the correct card-body class for the newest version of
Fedora Bootstrap.

update class syntax for label/badge

The previous version of fedora bootstrap used the label
class for badges/labels. This updates these
to the correct badge class for the newest version of
Fedora Bootstrap.

remove markdown blockquote style in site.css

This is now in fedora-markdown, so removing it
from the site.css

Related: #2333

make top menu aligned to the right

tweak spacing on elements to match how it was

make footer text colours reabable, and center text

remove box-shadow on top dropdown searchbar

Fix the text color of warning badges (yellow)

Previously the change to the new bootstrap caused
the warning badges (the yellow ones) to be a brighter
yellow / orange. This forces the badge colour back
to the old yellow / orange, and makes the text of
all these badges to be white as well.

Remove white bar at top of update page

make topbar dropdowns show a pointer

make revoke and unpush buttons a pointer

Signed-off-by: Ryan Lerch <rlerch@redhat.com>