Vagrant: Fold bdiff-cover into btest.
Having a separate bdiff-cover command is not useful since running
the tests are necessary to measure the test coverage, and since
diff-cover itself takes very little time to run in comparison to
the tests.

Therefore, it makes more sense to fold the diff-cover test into
btest and get rid of bdiff-cover.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>