Revert "Gracefully handle private bugs"
This reverts commit be369636529f3d637f9888bf658d77adac2e9b8b,
except for its migration. The migration must be kept since Bodhi
instances that have upgraded to 3.13.0 have this migration.

This commit used an attribute, private, in the Bugzilla API that
did not exist, which resulted in tracebacks in Bodhi's code that
alters or comments on Bugzilla tickets. This caused a production
outage for Fedora, so the quickest solution is to revert this
commit and fix it a different way.

fixes #3016

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>