Mark the bugs.private field as nullable.
In my fix for #3016, I opted to revert the commit that added the
private column to the bugs table, but not to remove the column
itself since that would be a backwards incompatible change.

Unfortunately, it did not occur to me that the private field being
non-nullable would mean that users would not be able to create
updates that have bugs associated with them, because there is no
longer code to set that column to a value.

This commit adjusts the column to be nullable so that we don't
introduce a backwards incompatible change, but also can allow users
to associate updates with bugs. Bodhi 4.0.0 will drop the column
entirely.

fixes #3024

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>