Only gate packages when Greenwave says 200 NOGO.
Greenwave has been found to have a response error rate of about 1
in 14 to 16 requests. This is causing Bodhi to gate quite a few
packages which is blocking updates from making it to stable.

This commit adjusts Bodhi so that it will mark the test gating
status as ignored if Greenwave responds with an HTTP 5xx code, or
if the request to Greenwave times out.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>