Add autopush and stable_days to an update.
This commit adds an autopush boolean and stable_days integer
to an update. When this boolean is true the approve_testing
cronjob will push the update to stable if it has meet the
testing requirements (stable_days).

stable_days cannot be smaller than the release
mandatory_days_in_testing.

Signed-off-by: Clement Verna <cverna@tutanota.com>

Use only stable approval comment.

This commit simplify how bodhi communicates that an update
is ready to be pushed to stable. Instead of including the reason
why an update is ready to be pushed (karma or time) bodhi
just informs that the update can now be pushed.

Signed-off-by: Clement Verna <cverna@tutanota.com>