Fix a flaky test.
Under high load, this particular test would fail because it depends
on the wall clock. The test can take long enough that arrow doesn't
return "just now", but something else instead. Thus, this patch
adjusts the test to mock arrow to return a particular string
instead, and asserts that a close enough value was passed to get().

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>