Use a dockerignore file to speed up CI.
This greatly reduces the size of the build context that we send to
the Docker daemon and helps the builds to go faster, especially
when they are able to cache the early layers.

I needed to slightly alter bodhi-ci as well to get it to bind mount
the .git folder in for the diff-cover test to work.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>