Do not warm up the home page cache in CI.
The home page cache takes a lot of time to build, and running this
in CI takes a lot of time as a result. This commit configures Bodhi
not to warm the cache for the home page upon process start, and as
a result the integration tests run about 58% faster on my laptop.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>