Do not crash CI when killing a non-running container.
Sometimes the DB container isn't running, and when the CI tests go
to kill it that raises an exception of its own. When this happens,
we lose the original traceback and the system is harder to debug.

This commit adjusts it so we don't crash if we try to kill a
container that isn't running.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>